RAG/requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
requests>=2.31.0
rich>=13.0.0
numpy>=1.24.0

